% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dfo_fields.R
\name{clean_dfo_fields}
\alias{clean_dfo_fields}
\title{clean_dfo_fields}
\usage{
clean_dfo_fields(df = NULL, customCrap = NULL)
}
\arguments{
\item{df}{default is \code{NULL}. This is the dataframe you want to drop
fields from.}

\item{customCrap}{default is \code{NULL}. This is a vector of fields you 
want removed from your dataframe.}
}
\value{
returns a data.frame without the columns CUSER, CDATE, UUSER, UDATE,
or any of the ones specified in customCrap.
}
\description{
This function facilitates the dropping of fields from a dataframe.
It was built for specific fields in MARFIS that I never wanted (i.e. CUSER, 
CDATE, UUSER, UDATE), but can also drop other fields easily.
}
\seealso{
Other data_manipulation:
\code{\link[=DDMMSS_to_DD]{DDMMSS_to_DD()}},
\code{\link[=DDMMx_to_DD]{DDMMx_to_DD()}},
\code{\link[=ISSETPROFILE_enwidener]{ISSETPROFILE_enwidener()}},
\code{\link[=dets_defuddler]{dets_defuddler()}},
\code{\link[=drop_cols]{drop_cols()}},
\code{\link[=rename_fields]{rename_fields()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{data_manipulation}
