% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2other_poly.R
\name{csv2other_poly}
\alias{csv2other_poly}
\title{csv2other_poly}
\usage{
csv2other_poly(csv = NULL, out = "sf", path = NULL)
}
\arguments{
\item{csv}{a PBSMapping PolySet to be converted.}

\item{out}{default is \code{'sf'}.  This indicates whether an \code{sf} object, 
\code{'sp'} (SpatialPolygonsDataFrame) or \code{'shp'} (shapefile) should be created.}

\item{path}{if \code{out ='shp'}, this is the path to where the shapefile 
should be saved.  sf and sp objects are just created in the environment}
}
\value{
an sf object, a SpatialPolygonsDataFrame,or a shapefile
}
\description{
This function will turn PBSMapping style csv into either an 
sf object, an sp object or a shapefile.
}
\seealso{
Other general_use: 
\code{\link{clip_by_poly_generic}()},
\code{\link{combine_lists}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_to_shp}()},
\code{\link{df_to_sp}()},
\code{\link{drop_cols}()},
\code{\link{pbs2sp_poly}()},
\code{\link{prepare_shape_fields}()},
\code{\link{updateCheck}()},
\code{\link{where_now}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
