% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDirValet.R
\name{dataDirValet}
\alias{dataDirValet}
\title{dataDirValet}
\usage{
dataDirValet(data.dir = NULL)
}
\arguments{
\item{data.dir}{The default is a "data" folder within your working directory. This is the path to where you want the
extracted files to go.}

\item{db}{default is \code{"ALL"}. This identifies the dataset you are working 
with.}
}
\description{
This cleans up the data.dir folder for use by various Mar.* packages
}
\note{
data is not added to the global environment by this function - changes are made, saved, and
dropped.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
