% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDMMSS_to_DD.R
\name{DDMMSS_to_DD}
\alias{DDMMSS_to_DD}
\title{DDMMSS_to_DD}
\usage{
DDMMSS_to_DD(
  df = NULL,
  lat.field = "LATITUDE",
  lon.field = "LONGITUDE",
  WestHemisphere = T
)
}
\arguments{
\item{df}{a dataframe to be analyzed. If left \code{NULL}, a value for \code{db} should be provided}

\item{lat.field}{the default is \code{"LATITUDE"}. This is the name of the 
field holding latitude values (in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  This is the name of the 
field holding longitudevalues (in decimal degrees)}

\item{WestHemisphere}{the default is \code{TRUE}.  It true, this ensure that all returned 
Longitudes are negative.  If false, the original sign is retained.}
}
\value{
a data.frame
}
\description{
This is a utility that will replace the DDMMSS columns of a dataframe with the same
coordinates in decimal degrees.
}
\seealso{
Other data_manipulation:
\code{\link[=DDMMx_to_DD]{DDMMx_to_DD()}},
\code{\link[=ISSETPROFILE_enwidener]{ISSETPROFILE_enwidener()}},
\code{\link[=clean_dfo_fields]{clean_dfo_fields()}},
\code{\link[=dets_defuddler]{dets_defuddler()}},
\code{\link[=drop_cols]{drop_cols()}},
\code{\link[=rename_fields]{rename_fields()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{data_manipulation}
