% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecationCheck.R
\name{deprecationCheck}
\alias{deprecationCheck}
\title{Check for Deprecated Parameters}
\usage{
deprecationCheck(...)
}
\arguments{
\item{...}{Parameters to check for deprecation.}
}
\description{
This function checks for deprecated parameters and issues a 
warning if any are found.
}
\seealso{
Other dfo_extractions:
\code{\link[=connectionCheck]{connectionCheck()}},
\code{\link[=get_data_tables]{get_data_tables()}},
\code{\link[=make_oracle_cxn]{make_oracle_cxn()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}l
}
\concept{dfo_extractions}
