% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETS_COL_LOOKUP.R
\docType{data}
\name{DETS_COL_LOOKUP}
\alias{DETS_COL_LOOKUP}
\title{DETS_COL_LOOKUP Metadata}
\format{
data.frame
\describe{
\item{COLUMN_DEFN_ID}{polygon identifier}
\item{SUM_DOC_DEFN_COL_ID}{polygon identifier}
\item{PROMPT_ENG}{polygon identifier}
}
}
\usage{
DETS_COL_LOOKUP
}
\description{
This is a lookup table used for coverting the codes used in the "long" marfis tables
into column names for the "wide" versions of the tables.
}
\keyword{datasets}
