% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dets_defuddler.R
\name{dets_defuddler}
\alias{dets_defuddler}
\title{dets_defuddler}
\usage{
dets_defuddler(marfName = NULL, df = NULL)
}
\arguments{
\item{marfName}{default is \code{NULL}.  This is the name that describes the marfis data sent to 
\code{df}.  It MUST be one of:
\itemize{
 \item \code{"LOG_EFRT_ENTRD_DETS"}
 \item \code{"LOG_SPC_ENTRD_DETS"}
 \item \code{"MON_DOC_ENTRD_DETS"}
 \item \code{"SD_LOG_ENTRD_DETS"}
 \item \code{"SD_SLP_ENTRD_DETS"}
 \item \code{"SLIP_SPC_ENTRD_DETS"} 
 \item \code{"SUM_DOC_ENTR_DETS"}
 }}

\item{df}{default is \code{NULL}.  This is a data.frame corresponding with one of the following
marfis objects:
\itemize{
 \item \code{"LOG_EFRT_ENTRD_DETS"}
 \item \code{"LOG_SPC_ENTRD_DETS"}
 \item \code{"MON_DOC_ENTRD_DETS"}
 \item \code{"SD_LOG_ENTRD_DETS"}
 \item \code{"SD_SLP_ENTRD_DETS"}
 \item \code{"SLIP_SPC_ENTRD_DETS"} 
 \item \code{"SUM_DOC_ENTR_DETS"}
 }
The data must retain the formatting of the original Oracle objects (i.e. column names)}
}
\description{
This function decodes the various marfis *_DETS table from long to wide, associating 
all of the data available for each discrete record.
}
\seealso{
Other data_manipulation:
\code{\link[=DDMMSS_to_DD]{DDMMSS_to_DD()}},
\code{\link[=DDMMx_to_DD]{DDMMx_to_DD()}},
\code{\link[=ISSETPROFILE_enwidener]{ISSETPROFILE_enwidener()}},
\code{\link[=clean_dfo_fields]{clean_dfo_fields()}},
\code{\link[=drop_cols]{drop_cols()}},
\code{\link[=rename_fields]{rename_fields()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{data_manipulation}
