% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_sf.R
\name{df_to_sf}
\alias{df_to_sf}
\title{df_to_sf}
\usage{
df_to_sf(
  df = NULL,
  lat.field = "LATITUDE",
  lon.field = "LONGITUDE",
  primary.object.field = "PID",
  secondary.object.field = "SID",
  order.field = "POS",
  type = "polys",
  user.crs = 4326
)
}
\arguments{
\item{df}{the default is \code{NULL}.  This is a dataframe that has coordinates holding latitudes
and longtitudes in decimal degrees.}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values
(in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude
values (in decimal degrees)}

\item{primary.object.field}{the default is \code{"PID"}.  If polygons or lines are specified,
this field is used to identify discrete objects.}

\item{secondary.object.field}{the default is \code{"SID"}. Complex polygons need this additional 
identifier in order to create objects such as polygons with holes.}

\item{order.field}{the default is \code{"POS"}. This parameter specifies the field that controls 
the order points should be joined.  In order to specifiy a "hole", the order field for the c
oordinates specifying the hole should be descending (rather than ascending).}

\item{type}{the default is \code{"polys"}, but it can also be \code{"lines"} or \code{"points"}}

\item{user.crs}{the default is \code{4326}.  This is the coordinate system that should
be assigned to spatial object created from \code{df}.}
}
\value{
an sf object
}
\description{
This function will accept a dataframe and turn it into an sf object of class 
"POLYGON", "LINESTRING" or "POINTS".  It can handle PBSMapping polysets or any appropriately 
structured dataframe.
}
\note{
this function was modified from 
https://github.com/Mar-scal/Assessment_fns/blob/master/Maps/pbs_2_sf.R
}
\seealso{
Other spatial:
\code{\link[=clip_by_poly]{clip_by_poly()}},
\code{\link[=df_qc_spatial]{df_qc_spatial()}},
\code{\link[=df_sf_to_gpkg]{df_sf_to_gpkg()}},
\code{\link[=gpkglayers_to_shapefiles]{gpkglayers_to_shapefiles()}},
\code{\link[=identify_area]{identify_area()}},
\code{\link[=make_segments]{make_segments()}},
\code{\link[=make_segments_isdb]{make_segments_isdb()}},
\code{\link[=prepare_shape_fields]{prepare_shape_fields()}}
}
\author{
Dave Keith \email{Dave.Keith@dfo-mpo.gc.ca}, Freya Keyser \email{Freya.Keyser@dfo-mpo.gc.ca} and adapted by Mike McMahon \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{spatial}
