% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_shp.R
\name{df_to_shp}
\alias{df_to_shp}
\title{df_to_shp}
\usage{
df_to_shp(
  df = NULL,
  lat.field = "LATITUDE",
  lon.field = "LONGITUDE",
  filename = NULL
)
}
\arguments{
\item{df}{the default is \code{NULL} a dataframe or spatial object whose fieldnames are to be processed}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values 
(in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude 
values (in decimal degrees)}

\item{filename}{the default is \code{NULL}.  This is the name of the output shapefile.  If \code{NULL},
the shape will be called "newShape".}
}
\value{
NULL - writes shapefile to current working directory
}
\description{
Super quick and dirty function for turning a dataframe with coordinates into a 
shapefile.  Only parameters are the Lat and Lon fields. Assumes WGS 84.
}
\note{
This function is set to overwrite existing shapefiles with the same name.
}
\seealso{
Other general_use: 
\code{\link{clip_by_poly_generic}()},
\code{\link{combine_lists}()},
\code{\link{convert2poly}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_to_sp}()},
\code{\link{drop_cols}()},
\code{\link{pbs2sp_poly}()},
\code{\link{prepare_shape_fields}()},
\code{\link{updateCheck}()},
\code{\link{where_now}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
