% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_sp.r
\name{df_to_sp}
\alias{df_to_sp}
\title{df_to_sp}
\usage{
df_to_sp(
  df = NULL,
  lat.field = "LATITUDE",
  lon.field = "LONGITUDE",
  the.crs = "EPSG:4326",
  autoQC = TRUE
)
}
\arguments{
\item{df}{default is \code{NULL}. This is the dataframe you want to spatialize.}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values
(in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude
values (in decimal degrees)}

\item{the.crs}{the default is \code{EPSG:4326}.  This is the projection 
of the incoming data.  This is NOT what you want the projection to be.  
Assigning a projection, and reprojecting are different.}

\item{autoQC}{the default is \code{TRUE}.  This runs df_qc_spatial() 
automatically, ensuring that all coordinates are in the Northern hemsisphere
(i.e. Latitude between 0 and 90N) and the western hemisphere (i.e. Longitude
between -180 and 0W)}
}
\value{
spatialpointsdataframe
}
\description{
This function takes a dataframe and converts in to a spatialpointsdataframe
}
\seealso{
Other general_use: 
\code{\link{aggregator}()},
\code{\link{clip_by_poly_generic}()},
\code{\link{combine_lists}()},
\code{\link{convert2poly}()},
\code{\link{df2sf}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_to_shp}()},
\code{\link{drop_cols}()},
\code{\link{pbs2sp_poly}()},
\code{\link{prepare_shape_fields}()},
\code{\link{subset_by_time}()},
\code{\link{updateCheck}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
