% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{.get_machine_id}
\alias{.get_machine_id}
\title{.get_machine_id}
\usage{
.get_machine_id(user = NULL, host = NULL)
}
\arguments{
\item{user}{Optional username override}

\item{host}{Optional hostname override}
}
\description{
supports encryption for data files
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\keyword{internal}
