% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{.save_encrypted}
\alias{.save_encrypted}
\title{.save_encrypted}
\usage{
.save_encrypted(object, file)
}
\arguments{
\item{object}{R object to save}

\item{file}{Path to save to}
}
\description{
saves data files such that they can't be easily loaded outside of the package
}
\keyword{internal}
