% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_NA_cols.R
\name{drop_NA_cols}
\alias{drop_NA_cols}
\title{drop_NA_cols}
\usage{
drop_NA_cols(df = NULL, dropUninformative = F, keepers = NULL)
}
\arguments{
\item{df}{default is \code{NULL}. This is the dataframe you want to check}

\item{dropUninformative}{default is \code{F}.  This can be set to TRUE to drop
all columns that have a single unique value throughout the entuire dataframe.}

\item{keepers}{default is \code{NULL}.This is a vector of columns you 
definitely want to retain, even all values are identical}
}
\value{
dataframe
}
\description{
This function takes a dataframe and removes any columns that are
entirely populated with NAs
}
\seealso{
Other general_use: \code{\link{df_qc_spatial}},
  \code{\link{df_to_sp}}, \code{\link{pbs2sp_poly}},
  \code{\link{prepare_shape_fields}},
  \code{\link{updateCheck}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
