% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkglayers_to_shps.R
\name{export_gpkglayers_to_shapefiles}
\alias{export_gpkglayers_to_shapefiles}
\title{export_gpkglayers_to_shapefiles}
\usage{
export_gpkglayers_to_shapefiles(
  gpkg_path = NULL,
  output_dir = NULL,
  crs = NULL
)
}
\arguments{
\item{gpkg_path}{This is a gpkg file with 1 or more layers inside that should be 
extracted to shapefiles}

\item{output_dir}{This is a path where you would like the shapefiles output to}

\item{crs}{the default is \code{NULL}.  This is the coordinate system  you would 
like the output shapefiles to be converted to (e.g. \code{4326}).  If left NULL,
no transformation will be attempted}
}
\description{
This function converts the layers within a gpkg file into shapefiles
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
