% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_tables.R
\name{get_data_tables}
\alias{get_data_tables}
\title{get_data_tables}
\usage{
get_data_tables(
  schema = NULL,
  data.dir = file.path(getwd()),
  tables = NULL,
  rownum = NULL,
  cxn = NULL,
  checkOnly = FALSE,
  force.extract = FALSE,
  env = .GlobalEnv,
  fuzzyMatch = TRUE,
  quietly = TRUE,
  extract_user = NULL,
  extract_computer = NULL
)
}
\arguments{
\item{schema}{default is \code{NULL}. This is the schema you want to access 
a additional tables from. A value of "<NA>" will check the schema associated 
with the logged in user (i.e. the user that made the connection).  An "<NA>"
schema is often necessary for ISDB data extractions.}

\item{data.dir}{The default is your working directory. If you are hoping to load existing data,
this should point to a folder containing your rdata files. If you are extracting data,
the data will be saved here.}

\item{tables}{The default value is \code{NULL}.  This is a vector of table 
names you want to extract that exist in identified \code{schema} of the the database.}

\item{rownum}{The default value is \code{NULL}. This is an integer that can be used to limit the 
number  of records retrieved from the specified table(s).  If it is left as \code{NULL}, all records 
retrieved.  Otherwise, it will return the specified number of rows.}

\item{cxn}{A valid Oracle connection object. This parameter allows you to 
pass an existing connection, reducing the need to establish a new connection 
within the function. If provided, it takes precedence over the connection-
related parameters.}

\item{checkOnly}{default is \code{FALSE}  This flag allows the function to be 
run such that it checks for the existence of the files, but doesn't load them.}

\item{force.extract}{default is \code{FALSE}  This flag forces a re-extraction of all of the 
tables (rather than loading previously extracted versions from data.dir)}

\item{env}{This the the environment you want this function to work in.  The 
default value is \code{.GlobalEnv}.}

\item{fuzzyMatch}{default is \code{TRUE}.  This allows source data tables to match with
(generally) synonymous  schema identifiers (i.e. MARFIS.==MARFISSCI.; RV.==GROUNDFISH.;
ISDB.=OBSERVER.)  Changing to False forces an exact match.}

\item{quietly}{default is \code{FALSE}.  If TRUE, no output messages will be shown.}

\item{extract_user}{default is \code{NULL}.  This parameter can be used with
\code{extract_computer} to load encypted data files extracted by another user
and/or computer}

\item{extract_computer}{default is \code{NULL}.  This parameter can be used with
\code{extract_user} to load encypted data files extracted by another user
and/or computer}
}
\description{
This function allows users to extract tables for which they have select permissions
from Oracle.
}
\seealso{
Other dfo_extractions:
\code{\link[=connectionCheck]{connectionCheck()}},
\code{\link[=deprecationCheck]{deprecationCheck()}},
\code{\link[=make_oracle_cxn]{make_oracle_cxn()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{dfo_extractions}
