% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc_hex.r
\docType{data}
\name{hex}
\alias{hex}
\title{This is the contents of the hex rdata file}
\description{
This is the contents of the hex rdata file
}
\references{
\url{data_blah.com}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\keyword{data}
