% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_area.R
\name{identify_area}
\alias{identify_area}
\title{identify_area}
\usage{
identify_area(
  df = NULL,
  lat.field = "LATITUDE",
  lon.field = "LONGITUDE",
  agg.poly.shp = NULL,
  agg.poly.field = NULL,
  flag.land = FALSE
)
}
\arguments{
\item{df}{a dataframe to be analyzed.}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values 
(in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude 
values (in decimal degrees)}

\item{agg.poly.shp}{default is \code{NULL}.  This is either the path to the 
*.shp file of a shapefile, an sf spatial object; or an sp spatialpolygonsdataframe.
 If NULL, NAFO zones will be used.}

\item{agg.poly.field}{default is \code{NULL}.  This identifies the field within 
\code{agg.poly.shp} that contains the values that should be appended to the 
input dataframe. If NULL, "NAFO_BEST", will be used, which is the finest
resolution NAFO subdivision.}

\item{flag.land}{default is \code{TRUE}.  This will result in positions determined to be on 
land to return the value "<LAND>" in the resulting dataframe.  
but would require updating}
}
\value{
a DataFrame with the column \code{agg.poly.field} added
}
\description{
This function takes a dataframe (with coordinate fields in decimal
degrees), (optionally) a spatial object (either a shapefile, an
sp::SpatialPolygonsDataframe, or an sf::polygon), and a field name from the 
polygon object.  It then overlays the df with the polygon, and determines which
discrete polygon contains each point from the df, and appends a new field 
containing this value.  
If no polygon is provided, the df will be assessed against NAFO subdivisions.
}
\seealso{
Other spatial:
\code{\link[=clip_by_poly]{clip_by_poly()}},
\code{\link[=df_qc_spatial]{df_qc_spatial()}},
\code{\link[=df_sf_to_gpkg]{df_sf_to_gpkg()}},
\code{\link[=df_to_sf]{df_to_sf()}},
\code{\link[=gpkglayers_to_shapefiles]{gpkglayers_to_shapefiles()}},
\code{\link[=make_segments]{make_segments()}},
\code{\link[=make_segments_isdb]{make_segments_isdb()}},
\code{\link[=prepare_shape_fields]{prepare_shape_fields()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{spatial}
