% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISSETPROFILE_enwidener.R
\name{ISSETPROFILE_enwidener}
\alias{ISSETPROFILE_enwidener}
\title{ISSETPROFILE_enwidener}
\usage{
ISSETPROFILE_enwidener(df)
}
\arguments{
\item{df}{A data.frame of the structure of ISDB.ISSETPROFILE - having 
columns SETDATE, SETTIME, LATITUDE, LONGITUDE, DEPTH, VESSEL_SPEED, AIR_TEMPERATURE, 
NET_TEMPERATURE, WATER_TEMPERATURE, BAR_PRESSURE, FISHSET_ID, SET_NO, PNTCD_ID.}
}
\value{
A dataframe.
}
\description{
This function reshapes the default ISDB.ISSETPROFILE data such that 
each fishing set gets a single record instead of being spread across 4 records. 
Values such as AIR_TMP which would have previously been associated with P1 - P4 
now become AIR_TMP1 through AIR_TMP4.  Additionally, LATITUDE, LONGITUDE and YEAR
fields are added using the first value for each of P1-P4 for each set.
}
\seealso{
Other data_manipulation:
\code{\link[=DDMMSS_to_DD]{DDMMSS_to_DD()}},
\code{\link[=DDMMx_to_DD]{DDMMx_to_DD()}},
\code{\link[=clean_dfo_fields]{clean_dfo_fields()}},
\code{\link[=dets_defuddler]{dets_defuddler()}},
\code{\link[=drop_cols]{drop_cols()}},
\code{\link[=rename_fields]{rename_fields()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{data_manipulation}
