% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_oracle_cxn.r
\name{make_oracle_cxn}
\alias{make_oracle_cxn}
\title{make_oracle_cxn}
\usage{
make_oracle_cxn(
  usepkg = "rodbc",
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  quietly = FALSE
)
}
\arguments{
\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username} 
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}  
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for 
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file), 
this can be left and that value will be used.  If a value for this is 
provided, it will take priority over your existing value.}

\item{quietly}{default is \code{FALSE}  This indicates whether or not status messages should be 
shown.}
}
\description{
This function facilitates creating a connection to Oracle, and
allows connection via RODBC or ROracle, depending on the value of usepkg.
}
\note{
This function temporarily changes the system values for "TZ" and "ORA_SDTZ" to "GMT" to 
ensure that dates and times extracted from the various databases remain correct.  It does this 
by running \code{Sys.setenv(TZ = "GMT")}  and \code{Sys.setenv(ORA_SDTZ = "GMT")}.  Once a 
connection is established (or failes to be established), it restores the values that previously 
existed.
}
\seealso{
Other dfo_extractions:
\code{\link[=connectionCheck]{connectionCheck()}},
\code{\link[=deprecationCheck]{deprecationCheck()}},
\code{\link[=get_data_tables]{get_data_tables()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{dfo_extractions}
