% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_isdb_tracks.r
\name{make_qc_isdb_tracks}
\alias{make_qc_isdb_tracks}
\title{make_qc_isdb_tracks}
\usage{
make_qc_isdb_tracks(isdb.df, return.choice = "lines")
}
\arguments{
\item{isdb.df}{This is the dataframe you want to plot.  isdb dataframes are
characterized by the existence of the following fields:
\itemize{
 \item \code{FISHSET_ID}
 \item \code{LAT1} and \code{LONG1}
 \item \code{LAT2} and \code{LONG2}
 \item \code{LAT3} and \code{LONG3}
 \item \code{LAT4} and \code{LONG4}
}}

\item{return.choice}{This determines what the function returns.  The default
value of \code{"lines"} returns a SpatialLinesDataFrame, where the dataframe 
contains the generated QC information.  Any other value for 
\code{return.choice} will instead return of data frame of all of the 
FISHSET_IDs, the QC determination of each line, the length of each line (in 
km), and the number of valid vertices. 
The QC data checks the positions of each FISHSET for NAs, Zeroes, or 
otherwise data.  Additionally, if the line has a length of zero, that is 
noted as well.}
}
\description{
This function takes a df of the isdb data (including fields
\code{FISHSET_ID} and LAT1:LAT4 and LONG1:LONG4) and makes a 
spatialLinesDataFrame which can be plotted.
Each line will have the following columns in the resulting data frame:
\itemize{
 \item \code{FISHSET_ID} This uniquely identifies a set
 \item \code{QC} This field identifies potential issues with the line - 
 including NA positions, positions that are "0", incorrect hemisphere, or 
 impossible coordinates
 \item \code{LEN_KM} This field shows the calculated distance of the 
 resultant line in kms
}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
