% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_segments.R
\name{make_segments}
\alias{make_segments}
\title{make_segments}
\usage{
make_segments(
  df,
  objField = "SEGMID",
  seqField = "POSITION_UTC_DATE",
  lat.field = "LATITUDE",
  lon.field = "LONGITUDE",
  points = "orphans",
  the.crs = "EPSG:4326",
  filename = NULL,
  create.spatial = TRUE,
  gpkgName = "make_segments.gpkg",
  path = NULL
)
}
\arguments{
\item{df}{default is \code{NULL}.  This is the dataframe to be processed.  It 
should have coordinates in decimal degrees.  It should also have fields for use by 
\code{objField} and \code{seqField}.}

\item{objField}{default is \code{NULL}. This is a field identifying which 
points are part of the same track.}

\item{seqField}{default is \code{NULL}. This is a field which can be used to 
correctly order the positions along the track (dates are fine).}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values 
(in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude 
values (in decimal degrees)}

\item{points}{default is \code{"orphans"}. Valid options include "orphans", "all" and "none". 
While this function primarily generates linestrings/tracks, it can also include the same data as 
points/vertices.  "none" will not include any vertices, "all" will include all vertices, and 
"orphans" will only include lone vertices, which are not part of any segment.}

\item{the.crs}{default is \code{"EPSG:4326"}. This is the projection 
you want any generated data to be output in.  Input data is assumed to be 
from a GPS and should be WGS84 (which is what the default value corresponds 
with).}

\item{filename}{default is \code{NULL}.  If you are outputting spatial objects, 
you can specify a name for them here.  This used to correspond with the output shapefile name,
but now is used to identify the layers within a gpkg file. They will also get a timestamp.}

\item{create.spatial}{default is \code{TRUE}.  This indicates whether or not to create a gpkg 
file in your working directory.}

\item{gpkgName}{default is \code{"make_segments.gpkg"}.  If \code{create.spatial = TRUE}, a gpkg 
file will be created, and the name here will control what it is called.}

\item{path}{this is the path to the gpkg file (e.g. "c:/folder/")}
}
\value{
a list containing sf objects.  Additionally, a gpkg can be generated.
}
\description{
This function takes a dataframe with coordinates in decimal 
degrees and a track identifier, and creates line segments for each distinct 
identifier.  If specified, it will generate spatial layers within a gpkg file.
}
\seealso{
Other spatial:
\code{\link[=clip_by_poly]{clip_by_poly()}},
\code{\link[=df_qc_spatial]{df_qc_spatial()}},
\code{\link[=df_sf_to_gpkg]{df_sf_to_gpkg()}},
\code{\link[=df_to_sf]{df_to_sf()}},
\code{\link[=gpkglayers_to_shapefiles]{gpkglayers_to_shapefiles()}},
\code{\link[=identify_area]{identify_area()}},
\code{\link[=make_segments_isdb]{make_segments_isdb()}},
\code{\link[=prepare_shape_fields]{prepare_shape_fields()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{spatial}
