% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSegments.R
\name{makeSegments}
\alias{makeSegments}
\title{makeSegments}
\usage{
makeSegments(df, objField = "SEGMID", seqField = "POSITION_UTC_DATE",
  points = "orphans", the.crs = "+init=epsg:4326", filename = NULL,
  plot = TRUE, createShp = TRUE)
}
\arguments{
\item{df}{default is \code{NULL}.  This is the dataframe to be processed.  It 
should have coordinates in decimal degrees and they should be in fields 
called "LATITUDE" and "LONGITUDE".  It should also have fields for use by 
\code{objField} and \code{seqField}.}

\item{objField}{default is \code{NULL}. This is a field identifying which 
points are part of the same track;}

\item{seqField}{default is \code{NULL}. This is a field which can be used to 
correctly order the positions along the track (dates are fine).}

\item{points}{default is \code{"orphans"}. While this function generates 
tracks, it's possible that single records can exist from which no track can 
be displayed.  Setting it to "orphans" includes these lone positions in the 
output plots and/or shapefiles.  Setting it to "all" includes all positions
in the output plots and/or shapefiles. Setting it to "none" ignores points
and doesn't output any.}

\item{the.crs}{default is \code{"+init=epsg:4326"}. This is the projection 
you want any generated data to be output in.  Input data is assumed to be 
from a GPS and should be WGS84 (which is what the default value corresponds 
with).}

\item{filename}{default is \code{NULL}.  If you are outputting shapefiles, 
you can specify a name for them here.  They will also get a timestamp.}

\item{plot}{default is \code{TRUE}. This determines whether or not generated 
tracks will be plotted.}

\item{createShp}{default is \code{TRUE}. This determines whether or not 
shapefiles will be generated in your working directory.}
}
\value{
a list with 2 items - a SpatialPointsDataFrame, and a 
SpatialLinesDataFrame.  Additionally, shapefiles can also be generated.
}
\description{
This function takes a dataframe with coordinates in decimal 
degrees and a track identifier, and creates line segments for each distinct 
identifier.  If specified, it will plot the results in R and/or create 
shapefiles.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
