% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector}

\item{na.rm}{The default value is \code{FALSE}.}
}
\description{
This function returns the mode of a vector, preserving data type.
If several values share the mode, all will be returned.
}
\note{
stolen from https://stackoverflow.com/questions/2547402/how-to-find-the-statistical-mode
}
\seealso{
Other util:
\code{\link[=SQL_in]{SQL_in()}},
\code{\link[=big_in]{big_in()}},
\code{\link[=combine_lists]{combine_lists()}},
\code{\link[=st_err]{st_err()}},
\code{\link[=updateCheck]{updateCheck()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{util}
