% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_zero.r
\name{na_zero}
\alias{na_zero}
\title{na_zero}
\usage{
na_zero(x)
}
\arguments{
\item{x}{The default value is \code{NULL}.}
}
\description{
Converts all NAs to 0s
}
\seealso{
Other util: \code{\link{SQL_in}}, \code{\link{st_err}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}l
}
