% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs2sp_poly.R
\name{pbs2sp_poly}
\alias{pbs2sp_poly}
\title{pbs2sp_poly}
\usage{
pbs2sp_poly(pbs = NULL, create.shp = FALSE)
}
\arguments{
\item{pbs}{a PBSMapping PolySet to be converted.}

\item{create.shp}{default is \code{FALSE}.  This indicates whether or not 
a shapefile should be created.}
}
\value{
a SpatialPolygonsDataFrame, and potentially a shapefile
}
\description{
This function will turn PBSMapping PolySets into a 
SpatialPolygonsDataFrame, and can generate a shapefile at teh same time, if 
requested.
}
\seealso{
Other general_use: 
\code{\link{clip_by_poly_generic}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_to_shp}()},
\code{\link{df_to_sp}()},
\code{\link{drop_cols}()},
\code{\link{prepare_shape_fields}()},
\code{\link{updateCheck}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
