% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hex_data.R
\name{plot_hex_data}
\alias{plot_hex_data}
\title{plot_hex_data}
\usage{
plot_hex_data(
  data_sf = NULL,
  plotfld = NULL,
  extent = "data",
  hideEmptyCells = TRUE
)
}
\arguments{
\item{data_sf}{default is \code{NULL}.  an sf version of the Grid2Min object from assess_privacy.}

\item{plotfld}{default is \code{NULL}.  This is a single field within the data object that you 
want to plot. Leaving it NULL results in a popup box of selectable fields.}

\item{extent}{default is \code{"data"}.  This controls the extent of the output plot.  It can be one 
of \code{"data"}, \code{"layer"}, or a vector of coordinates corresponding with 
xmin, xmax, ymin, ymax (e.g. \code{bbox = c(-65,-63,43,45)}.
\itemize{
 \item \code{"data"} - results in the plot being restricted to an area where values of \code{plotfld} are 
non-zero
  \item \code{"layer"} - results in the plot showing the full extent of the gridded area
  \item \code{c(xmin, xmax, ymin, ymax)} - results in a plot restricted to the specified area
  }}

\item{hideEmptyCells}{default is \code{TRUE}.  By default, this parameter results in the plot 
only showing grid cells for which data exists (i.e. non-zero).  This offers a dramatic performance 
improvement.  If the entire mesh of predominantly empty cells must be shown, set this to FALSE.}
}
\value{
an ggplot2 plot
}
\description{
This function plots the gridded data from assess_privacy
}
\note{
if no  PID or POS is provided, it is assumed that the provided positions
are for a single polygon and the position are in the correct order.
}
\seealso{
Other privacy:
\code{\link[=aggregator]{aggregator()}},
\code{\link[=assess_privacy]{assess_privacy()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{privacy}
