% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_in_poly_stats.R
\name{point_in_poly_stats}
\alias{point_in_poly_stats}
\title{point_in_poly_stats}
\usage{
point_in_poly_stats(
  sf_points,
  sf_points_field = NULL,
  sf_polys,
  sf_polys_field = NULL
)
}
\arguments{
\item{sf_points}{This is an sf point object}

\item{sf_points_field}{This is the field within \code{sf_points} for which you would like to calculate a mean and standard error (for a given polygon)}

\item{sf_polys}{This is an sf polygon object}

\item{sf_polys_field}{This is the field from \code{sf_polys} that identifies the polygons (e.g. a unique identifier for each polygon)}
}
\value{
an sf polygon
}
\description{
This function takes an sf point object, an sf polygon object, and a 
 fieldname from each and outputs the sf polygon object with additional fields 
 showing the count of the sf_points within each polygon, as well as the mean 
 and standard error for a specified field for a specified field
}
