% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shape_fields.r
\name{prepare_shape_fields}
\alias{prepare_shape_fields}
\title{prepare_shape_fields}
\usage{
prepare_shape_fields(shape)
}
\arguments{
\item{shape}{a dataframe or spatial object whose fieldnames are to be processed}
}
\value{
a dataframe - spatial or otherwise
}
\description{
Shapefiles are limited in the naming conventions of the fields - they can only
be 10 characters, and must be devoid of special characters that are allowed in R.  This function
attempts to ensure that the shortened names that result from calls to writeOGR() can still be 
understood and are not simply truncated.
}
\seealso{
Other spatial:
\code{\link[=clip_by_poly]{clip_by_poly()}},
\code{\link[=df_qc_spatial]{df_qc_spatial()}},
\code{\link[=df_sf_to_gpkg]{df_sf_to_gpkg()}},
\code{\link[=df_to_sf]{df_to_sf()}},
\code{\link[=gpkglayers_to_shapefiles]{gpkglayers_to_shapefiles()}},
\code{\link[=identify_area]{identify_area()}},
\code{\link[=make_segments]{make_segments()}},
\code{\link[=make_segments_isdb]{make_segments_isdb()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{spatial}
