% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_date.R
\name{simple_date}
\alias{simple_date}
\title{simple_date}
\usage{
simple_date(df = NULL, datefields = NULL)
}
\arguments{
\item{df}{default is \code{NULL}. This is the dataframe with datetime fields 
you want to modify.}

\item{datefields}{default is \code{NULL}. This is a vector of datetime fields
that you want to overwrite with simpler, date fields (i.e. 'YYYY-MM-DD').}
}
\value{
returns a data.frame with the updated fields.
}
\description{
This function converts specified datetime fields into date fields.
}
\note{
The number of items in oldFields and newFields must be identical, and 
the order matters in that newFields[3] will replace oldFields[3].
}
\seealso{
Other datesAndTimes:
\code{\link[=vali_dates]{vali_dates()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{datesAndTimes}
