% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_now.R
\name{where_now}
\alias{where_now}
\title{where_now}
\usage{
where_now(returnTime = FALSE, callstack = sys.calls())
}
\arguments{
\item{returnTime}{default is \code{FALSE}.  If TRUE, this function will return a value for 
proc.time, facilitating timing of functions}

\item{callstack}{default is \code{sys.calls()}.  Not meant to be changed.}
}
\value{
nothing, it just writes messages to the screen
}
\description{
**DEPRECATED:** This function indicates which function has been reached.

**This function is deprecated and will be removed in a future version.** 
Please use \code{\link{debugMode}} instead, which provides similar functionality without 
requiring code changes to your functions.

To replace \code{where_now()} calls:
\itemize{
  \item Remove all \code{where_now()} calls from your functions
  \item Use \code{debugMode(TRUE)} to enable debugging for all package functions
  \item Use \code{debugMode(FALSE)} to disable debugging
}

The new approach provides automatic entry/exit messages with timing for all functions,
eliminating the need to manually add debugging code.
}
\seealso{
\code{\link{debugMode}} for the replacement debugging approach

Other debugging:
\code{\link[=changeDetector]{changeDetector()}},
\code{\link[=updateExpected]{updateExpected()}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{debugging}
