% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeDetector.R
\name{changeDetector}
\alias{changeDetector}
\title{changeDetector}
\usage{
changeDetector(pre_ = NULL, post_ = NULL, fields = NULL, flagTxt = NULL)
}
\arguments{
\item{pre_}{default is \code{NA} This is dataframe prior to the change}

\item{post_}{default  is \code{NULL} TThis is dataframe after the change}

\item{fields}{default  is \code{NULL} This is a vector of field names for which you are interested in the number of unique values before and after the change}

\item{flagTxt}{default is \code{NULL} This is text you would like to be displayed before listing the results.  For debugging, it's useful to set this text as 
a description of the processing step that was just done.}
}
\value{
nothing, it just writes messages to the screen
}
\description{
This function compares a dataframe before and after something has been done to it.  It will indicate the number of rows (before and after), and 
if \code{fields} are specified, it will also indicate how many unique values of each of the specified fields were present before and after the change.
}
\seealso{
Other debugging: 
\code{\link{updateExpected}()},
\code{\link{where_now}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{debugging}
