% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_lists.R
\name{combine_lists}
\alias{combine_lists}
\title{combine_lists}
\usage{
combine_lists(primary = NULL, ancilliary = NULL, quiet = T)
}
\arguments{
\item{primary}{default is \code{NULL}}

\item{ancilliary}{default is \code{NULL}}

\item{quiet}{default is \code{TRUE}.  If FALSE, this function will notify the
user that the lists have similarly named elements, and will indicate which
values will be retained (i.e. those from the primary list)}
}
\value{
list of all of the arguments
}
\description{
This function assembles 2 lists into a single list.  When
elements from both lists have the same name, the values from the primary list
will be retained.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
