% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_lists.R
\name{combine_lists}
\alias{combine_lists}
\title{combine_lists}
\usage{
combine_lists(primary = NULL, ancilliary = NULL, quietly = FALSE)
}
\arguments{
\item{primary}{default is \code{NULL}}

\item{ancilliary}{default is \code{NULL}}

\item{quietly}{default is \code{FALSE}.  By default, this function will indicate the names and values of duplicated (i.e. ancilliary) list elements that will be dropped}
}
\value{
list of all of the arguments
}
\description{
This function assembles 2 lists into a single list.  When
elements from both lists have the same name, the values from the primary list
will be retained.
}
\seealso{
Other general_use: 
\code{\link{clip_by_poly_generic}()},
\code{\link{convert2poly}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_to_shp}()},
\code{\link{df_to_sp}()},
\code{\link{drop_cols}()},
\code{\link{pbs2sp_poly}()},
\code{\link{prepare_shape_fields}()},
\code{\link{updateCheck}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
