% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionCheck.R
\name{connectionCheck}
\alias{connectionCheck}
\title{Check Connection Type}
\usage{
connectionCheck(cxn)
}
\arguments{
\item{cxn}{The connection object to check.}
}
\value{
The command function corresponding to the connection type.
}
\description{
This function checks the type of the provided connection and 
returns the appropriate command function.
}
\seealso{
Other dfo_extractions: 
\code{\link{deprecationCheck}()},
\code{\link{get_data_tables}()},
\code{\link{make_oracle_cxn}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}l
}
\concept{dfo_extractions}
