% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.dir_valet.R
\name{data.dir_valet}
\alias{data.dir_valet}
\title{data.dir_valet}
\usage{
data.dir_valet(data.dir = NULL)
}
\arguments{
\item{data.dir}{The default is a "data" folder within your working directory. This is the path to where you want the
extracted files to go.}
}
\description{
This cleans up the data.dir folder for use by various Mar.* packages
}
\note{
data is not added to the global environment by this function - changes are made, saved, and
dropped.
}
\seealso{
Other file_management: 
\code{\link{load_encrypted}()},
\code{\link{save_encrypted}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{file_management}
