% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDMMSS_to_DD.R
\name{DDMMSS_to_DD}
\alias{DDMMSS_to_DD}
\title{DDMMSS_to_DD}
\usage{
DDMMSS_to_DD(
  df = NULL,
  lat.field = "LATITUDE",
  lon.field = "LONGITUDE",
  WestHemisphere = T
)
}
\arguments{
\item{df}{a dataframe to be analyzed. If left \code{NULL}, a value for \code{db} should be provided}

\item{lat.field}{the default is \code{"LATITUDE"}. This is the name of the 
field holding latitude values (in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  This is the name of the 
field holding longitudevalues (in decimal degrees)}

\item{WestHemisphere}{the default is \code{TRUE}.  It true, this ensure that all returned 
Longitudes are negative.  If false, the original sign is retained.}
}
\value{
a data.frame
}
\description{
This is a utility that will replace the DDMMSS columns of a dataframe with the same
coordinates in decimal degrees.
}
\seealso{
Other general_use: 
\code{\link{aggregator}()},
\code{\link{clip_by_poly_generic}()},
\code{\link{combine_lists}()},
\code{\link{convert2poly}()},
\code{\link{df2sf}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_sf_to_gpkg}()},
\code{\link{df_to_sf}()},
\code{\link{df_to_shp}()},
\code{\link{df_to_sp}()},
\code{\link{drop_cols}()},
\code{\link{prepare_shape_fields}()},
\code{\link{subset_by_time}()},
\code{\link{updateCheck}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
