% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_qc_spatial.r
\name{df_qc_spatial}
\alias{df_qc_spatial}
\title{df_qc_spatial}
\usage{
df_qc_spatial(df, lat.field = "LATITUDE", lon.field = "LONGITUDE",
  return.bad = FALSE)
}
\arguments{
\item{df}{the dataframe to be cleaned}

\item{lat.field}{default is \code{LATITUDE}.  The field in the dataframe holding the latitudes}

\item{lon.field}{default is \code{LONGITUDE}.  The field in the dataframe holding the longitudes}

\item{return.bad}{the default is \code{FALSE}.  Normally, only the plottable points are returned.
If TRUE, only the rejected points are returned.}
}
\value{
data.frame contents depend of the return.bad parameter
}
\description{
This function drops points that can't be plotted from a data frame.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
