% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_qc_spatial.r
\name{df_qc_spatial}
\alias{df_qc_spatial}
\title{df_qc_spatial}
\usage{
df_qc_spatial(
  df,
  lat.field = "LATITUDE",
  lon.field = "LONGITUDE",
  return.bad = FALSE,
  autoQC = T
)
}
\arguments{
\item{df}{the dataframe to be cleaned}

\item{lat.field}{default is \code{LATITUDE}.  The field in the dataframe holding the latitudes}

\item{lon.field}{default is \code{LONGITUDE}.  The field in the dataframe holding the longitudes}

\item{return.bad}{the default is \code{FALSE}.  Normally, only the plottable points are returned.
If TRUE, only the rejected points are returned.}

\item{autoQC}{the default is \code{TRUE}.  This runs df_qc_spatial() 
automatically, ensuring that all coordinates are in the Northern hemsisphere
(i.e. Latitude between 0 and 90N) and the western hemisphere (i.e. Longitude
between -180 and 0W)}
}
\value{
data.frame contents depend of the return.bad parameter
}
\description{
This function identifies points that aren't in the northern 
hemisphere (i.e. LAT between 0 and 90) and aren't in the western hemisphere 
(i.e. LON between 0 and 180).  By default, it returns the "good" points, but 
the "return.bad" flag allows it to return only the "bad" points.
}
\seealso{
Other spatial: 
\code{\link{clip_by_poly}()},
\code{\link{df_sf_to_gpkg}()},
\code{\link{df_to_sf}()},
\code{\link{gpkglayers_to_shapefiles}()},
\code{\link{identify_area}()},
\code{\link{make_segments}()},
\code{\link{make_segments_isdb}()},
\code{\link{prepare_shape_fields}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{spatial}
