% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_sf_to_gpkg.R
\name{df_sf_to_gpkg}
\alias{df_sf_to_gpkg}
\title{df_sf_to_gpkg}
\usage{
df_sf_to_gpkg(
  df = NULL,
  lat.field = "LATITUDE",
  lon.field = "LONGITUDE",
  crs = 4326,
  makeSpatial = T,
  path = NULL,
  layerName = NULL,
  gpkgName = "r_layers.gpkg"
)
}
\arguments{
\item{df}{the default is \code{NULL} a dataframe or sf object to be added to a gpkg file}

\item{lat.field}{the default is \code{"LATITUDE"}. For dataframe objects, this is the name of the 
field holding latitude values (in decimal degrees).  This field can be ignored if the submitted 
df is an sf object.}

\item{lon.field}{the default is \code{"LONGITUDE"}. For dataframe objects, this is the name of 
the field holding longitude values (in decimal degrees).  This field can be ignored if the 
submitted df is an sf object.}

\item{crs}{the default is \code{4326}.  This is the coordinate system associated with the input
positions, and it assumes WGS84 (i.e. collected via a GPS). This field can be ignored if the 
submitted df is an sf object.}

\item{makeSpatial}{the default is \code{TRUE}. By default, this function will attempt to convert 
dataframes to spatial (i.e. sf) objects.  Set this to false if you just want tabular data.}

\item{path}{this is the path to the gpkg file (e.g. "c:/folder/")}

\item{layerName}{the default is \code{NULL}.  This will be the name of the layer within the gpkg 
file (not the name of the generated gpkg file.)}

\item{gpkgName}{the default is \code{"r_layers.gpkg"}.  This is the name of the gpkg file that 
will hold the spatial layers.  If this file already exists in \code{path}, the layer will be 
added to the existing file.  If the file does not exist, it will be created.}
}
\value{
NULL - writes shapefile to current working directory
}
\description{
This function adds dataframes (with coordinates) and/or sf spatial objects into a 
gpkg file. Dataframes will be converted into sf POINT objects, but if existing sf objects are 
submitted, they will be added to the gpkg file.
}
\seealso{
Other spatial: 
\code{\link{clip_by_poly}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_to_sf}()},
\code{\link{gpkglayers_to_shapefiles}()},
\code{\link{identify_area}()},
\code{\link{make_segments}()},
\code{\link{make_segments_isdb}()},
\code{\link{prepare_shape_fields}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{spatial}
