% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_cols.R
\name{drop_cols}
\alias{drop_cols}
\title{drop_cols}
\usage{
drop_cols(
  df = NULL,
  justDropNAs = TRUE,
  uniformFields = "retain",
  keepFields = NULL,
  quietly = T
)
}
\arguments{
\item{df}{default is \code{NULL}. This is the dataframe you want to handle.}

\item{justDropNAs}{default is \code{TRUE}.  If this value is TRUE, the script 
will only drop fields that are solely populated with NAs.  If FALSE, it
will drop/retain iother fields depending on the following parameters.}

\item{uniformFields}{default is \code{'retain'}.  This determines how the script
will handle uniform fields.  It can either retain them or drop them.  Valid
values are \code{'retain'} or \code{'drop'}.   
\code{'drop'} specifies that any field that has a single value for all 
entries will be dropped.  This is useful for identifying the fields that can 
be used to differentiate records.
\code{'retain'} specifies that only those fields which have a single 
value for all entries will be retained.  This is useful for identifying the 
fields with values that are shared by all records in a dataframe.}

\item{keepFields}{default is \code{NULL}.This is a vector of column names you 
definitely want to retain, regardless of the values.}

\item{quietly}{default is \code{FALSE}.  By default, this function will not output messages,
but they can be turned on by setting this to TRUE}
}
\value{
dataframe
}
\description{
This function takes a dataframe and drops or retains fields
based on the uniformity of their values.
}
\seealso{
Other general_use: 
\code{\link{aggregator}()},
\code{\link{clip_by_poly_generic}()},
\code{\link{combine_lists}()},
\code{\link{convert2poly}()},
\code{\link{df2sf}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_to_shp}()},
\code{\link{df_to_sp}()},
\code{\link{pbs2sp_poly}()},
\code{\link{prepare_shape_fields}()},
\code{\link{subset_by_time}()},
\code{\link{updateCheck}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
