% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_NA_cols.R
\name{drop_NA_cols}
\alias{drop_NA_cols}
\title{drop_NA_cols}
\usage{
drop_NA_cols(df)
}
\arguments{
\item{df}{default is \code{NULL}. This is the dataframe you want to check}
}
\value{
dataframe
}
\description{
This function takes a dataframe and removes any columns that are
entirely populated with NAs
}
\seealso{
Other general_use: \code{\link{df_qc_spatial}},
  \code{\link{df_to_sp}},
  \code{\link{prepare_shape_fields}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
