% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_tables.R
\name{get_data_tables}
\alias{get_data_tables}
\title{get_data_tables}
\usage{
get_data_tables(
  schema = NULL,
  data.dir = file.path(getwd(), "data"),
  tables = NULL,
  rownum = NULL,
  cxn = NULL,
  usepkg = "rodbc",
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  checkOnly = FALSE,
  force.extract = FALSE,
  env = .GlobalEnv,
  fuzzyMatch = TRUE,
  quietly = TRUE
)
}
\arguments{
\item{schema}{default is \code{NULL}. This is the schema you want to access 
a additional tables from.}

\item{data.dir}{The default is your working directory. If you are hoping to load existing data,
this folder should contain a data folder containing your rdata files. If you are extracting data,
a data folder will be created under this folder.
extracted files to go.}

\item{tables}{The default value is \code{NULL}.  This is a vector of table 
names you want to extract that exist in identified \code{schema} of the the database.}

\item{rownum}{The default value is \code{NULL}. This is an integer that can be used to limit the 
number  of records retrieved from the specified table(s).  If it is left as \code{NULL}, all records 
retrieved.  Otherwise, it will return the specified number of rows.}

\item{cxn}{A valid Oracle connection object. This parameter allows you to 
pass an existing connection, reducing the need to establish a new connection 
within the function. If provided, it takes precedence over the connection-
related parameters.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the 
connection to Oracle should use \code{'rodbc'} or \code{'roracle'} to 
connect.  rodbc is slightly easier to setup, but roracle will extract data ~ 
5x faster.}

\item{fn.oracle.username}{Default is \code{'_none_'}. This is your username 
for accessing Oracle objects. If you have a value for \code{oracle.username} 
stored in your environment (e.g., from an rprofile file), this can be left 
out and that value will be used. If a value for this is provided, it will 
take priority over your existing value. Deprecated; use \code{cxn} instead.}

\item{fn.oracle.password}{Default is \code{'_none_'}. This is your password 
for accessing Oracle objects. If you have a value for \code{oracle.password} 
stored in your environment (e.g., from an rprofile file), this can be left 
out and that value will be used. If a value for this is provided, it will 
take priority over your existing value. Deprecated; use \code{cxn} instead.}

\item{fn.oracle.dsn}{Default is \code{'_none_'}. This is your DSN/ODBC 
identifier for accessing Oracle objects. If you have a value 
for \code{oracle.dsn} stored in your environment (e.g., from an rprofile 
file), this can be left out and that value will be used. If a value for this 
is provided, it will take priority over your existing value. Deprecated; use 
\code{cxn} instead.}

\item{checkOnly}{default is \code{FALSE}  This flag allows the function to be 
run such that it checks for the existence of the files, but doesn't load them.}

\item{force.extract}{default is \code{FALSE}  This flag forces a re-extraction of all of the 
tables (rather than loading previously extracted versions from data.dir)}

\item{env}{This the the environment you want this function to work in.  The 
default value is \code{.GlobalEnv}.}

\item{fuzzyMatch}{default is \code{TRUE}.  This allows source data tables to match with
(generally) synonymous  schema identifiers (i.e. MARFIS.==MARFISSCI.; RV.==GROUNDFISH.;
ISDB.=OBSERVER.)  Changing to False forces an exact match.}

\item{quietly}{default is \code{FALSE}.  If TRUE, no output messages will be shown.}
}
\description{
This function allows users to extract tables for which they have select permissions
from Oracle.
}
\seealso{
Other dfo_extractions: 
\code{\link{make_oracle_cxn}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{dfo_extractions}
