% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_tables.R
\name{get_data_tables}
\alias{get_data_tables}
\title{get_data_tables}
\usage{
get_data_tables(
  schema = NULL,
  data.dir = file.path(getwd(), "data"),
  tables = NULL,
  usepkg = "rodbc",
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  env = .GlobalEnv,
  quietly = F
)
}
\arguments{
\item{schema}{default is \code{NULL}. This is the schema you want to access 
a additional tables from.}

\item{data.dir}{The default is your working directory. If you are hoping to load existing data,
this folder should contain a data folder containing your rdata files. If you are extracting data,
a data folder will be created under this folder.
extracted files to go.}

\item{tables}{The default value is \code{NULL}.  This is a vector of table 
names you want to extract that exist in identified \code{schema} of the the database.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the 
connection to Oracle should use \code{'rodbc'} or \code{'roracle'} to 
connect.  rodbc is slightly easier to setup, but roracle will extract data ~ 
5x faster.}

\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects.}

\item{env}{This the the environment you want this function to work in.  The 
default value is \code{.GlobalEnv}.}

\item{quietly}{default is \code{FALSE}.  If TRUE, no output messages will be shown.}
}
\description{
This function allows users to extract tables for which they have select permissions
from Oracle.
}
\seealso{
Other dfo_extractions: 
\code{\link{make_oracle_cxn}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{dfo_extractions}
