% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpkglayers_to_shps.R
\name{gpkglayers_to_shapefiles}
\alias{gpkglayers_to_shapefiles}
\title{gpkglayers_to_shapefiles}
\usage{
gpkglayers_to_shapefiles(gpkg_path = NULL, output_dir = NULL, crs = NULL)
}
\arguments{
\item{gpkg_path}{This is a gpkg file with 1 or more layers inside that should be 
extracted to shapefiles}

\item{output_dir}{This is a path where you would like the shapefiles output to}

\item{crs}{the default is \code{NULL}.  This is the coordinate system  you would 
like the output shapefiles to be converted to (e.g. \code{4326}).  If left NULL,
no transformation will be attempted}
}
\description{
This function converts the layers within a gpkg file into shapefiles
}
\seealso{
Other spatial: 
\code{\link{clip_by_poly}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_sf_to_gpkg}()},
\code{\link{df_to_sf}()},
\code{\link{identify_area}()},
\code{\link{make_segments}()},
\code{\link{make_segments_isdb}()},
\code{\link{prepare_shape_fields}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{spatial}
