% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_area.R
\name{identify_area}
\alias{identify_area}
\title{identify_area}
\usage{
identify_area(
  df = NULL,
  lat.field = "LATITUDE",
  lon.field = "LONGITUDE",
  agg.poly.shp = NULL,
  agg.poly.field = NULL
)
}
\arguments{
\item{df}{a dataframe to be analyzed. If left \code{NULL}, a value for 
\code{db} should be provided}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values 
(in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude 
values (in decimal degrees)}

\item{agg.poly.shp}{default is \code{NULL}.  This is the shapefile that has 
polygons that should be checked for sufficient unique values of the 
sens.fields.  If NULL, NAFO zones will be used.  Otherwise, a path to any 
polygon shapefile can be provided.}

\item{agg.poly.field}{default is \code{NULL}.  This identifies the field within 
the shapefile provided to agg.poly.shp that should be used to check for 
sufficient unique values of the sens.fields.}
}
\value{
a DataFrame with the column \code{agg.poly.field} added
}
\description{
This function takes a df with coordinate fields in decimal
degrees, and overlays it a shapefile (agg.poly.shp).  It adds a column to the 
df indicating which polygon within the shapefile each point falls within.
If no polygon is provided, the df will be assessed against NAFO subdivisions.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
