% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISSETPROFILE_enwidener.R
\name{ISSETPROFILE_enwidener}
\alias{ISSETPROFILE_enwidener}
\title{ISSETPROFILE_enwidener}
\usage{
ISSETPROFILE_enwidener(ISSETPROFILE)
}
\arguments{
\item{ISSETPROFILE}{A data.frame of the structure of ISDB.ISSETPROFILE - having 
columns SETDATE, SETTIME, LATITUDE, LONGITUDE, DEPTH, VESSEL_SPEED, AIR_TEMPERATURE, 
NET_TEMPERATURE, WATER_TEMPERATURE, BAR_PRESSURE, FISHSET_ID, SET_NO, PNTCD_ID.}
}
\value{
A data.table in the same shape as your SQL view.
}
\description{
This function reshapes the default ISDB.ISSETPROFILE data such that 
each fishing set gets a single record instead of being spread across 4 records. 
Values such as AIR_TMP which would have previously been associated with P1 - P4 
now become AIR_TMP1 through AIR_TMP4.  Additionally, LATITUDE, LONGITUDE and YEAR
fields are added using the first value for each of P1-P4 for each set.
}
