% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{load_encrypted}
\alias{load_encrypted}
\title{.load_encrypted}
\usage{
load_encrypted(
  file,
  extract_user = NULL,
  extract_computer = NULL,
  envir = parent.frame()
)
}
\arguments{
\item{file}{Path to encrypted file}

\item{extract_user}{Optional username of original file creator}

\item{extract_computer}{Optional hostname of original file creator}

\item{envir}{Environment where loaded objects will be assigned}
}
\description{
loads encrypted data files
}
