% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{load_encrypted}
\alias{load_encrypted}
\title{load_encrypted}
\usage{
load_encrypted(
  file,
  extract_user = NULL,
  extract_computer = NULL,
  envir = parent.frame()
)
}
\arguments{
\item{file}{Path to the file to load}

\item{extract_user}{Optional username of original file creator}

\item{extract_computer}{Optional hostname of original file creator}

\item{envir}{Environment where loaded objects will be assigned}
}
\value{
Names of loaded objects (invisibly)
}
\description{
Loads data files, handling both encrypted and unencrypted formats
}
\seealso{
Other file_management: 
\code{\link{data.dir_valet}()},
\code{\link{save_encrypted}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{file_management}
