% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_segments_isdb.r
\name{make_segments_isdb}
\alias{make_segments_isdb}
\title{make_segments_isdb}
\usage{
make_segments_isdb(isdb.df, filename = NULL, create.spatial = TRUE)
}
\arguments{
\item{isdb.df}{This is the dataframe you want to plot.  isdb dataframes are
characterized by the existence of one of the following sets of fields:
ISSETPROFILE_WIDE - based:
\itemize{
 \item \code{FISHSET_ID}
 \item \code{SET_NO}
 \item \code{LAT1}, \code{LONG1} and \code{DATE_TIME1}
 \item \code{LAT2}, \code{LONG2} and \code{DATE_TIME2}
 \item \code{LAT3}, \code{LONG3} and \code{DATE_TIME3}
 \item \code{LAT4}, \code{LONG4} and \code{DATE_TIME4}
}
or
ISSETPROFILE - based
\itemize{
\item \code{FISHSET_ID}
 \item \code{SET_NO}
 \item \code{LATITUDE}, \code{LONGITUDE}, \code{DATETIME} and \code{PNTCD_ID}
}}

\item{filename}{default is \code{NULL}.  If you are outputting shapefiles, 
you can specify a name for them here.  They will also get a timestamp.}

\item{create.spatial}{default is \code{TRUE}.  This indicates whether or not to create a gpkg 
file in your working directory.}
}
\description{
This function takes a df of the isdb data, and creates line segments for each 
distinct FISHSET_ID. If specified, it will generate spatial layers within a gpkg file.
Each line will have the following columns in the resulting data frame:
\itemize{
 \item \code{FISHSET_ID} This uniquely identifies a set
 \item \code{LEN_KM} This field shows the calculated distance of the 
 resultant line in kms
 \item \code{SET_NO} This uniquely identifies a set
 \item \code{N_VALID_VERT} This field shows how many vertices appear 
 correct after NAs, 0s and other problematic values have been dropped
 \item \code{SETSTART} This is the earliest time associated with a set
 \item \code{SETEND}  This is the last time associated with a set
 \item \code{SET_DUR} This is the difference in time (in hours) between the first and last 
 position of a set
 }
}
\seealso{
Other spatial: 
\code{\link{clip_by_poly}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_sf_to_gpkg}()},
\code{\link{df_to_sf}()},
\code{\link{gpkglayers_to_shapefiles}()},
\code{\link{identify_area}()},
\code{\link{make_segments}()},
\code{\link{prepare_shape_fields}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{spatial}
