% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_segments_isdb.r
\name{make_segments_isdb}
\alias{make_segments_isdb}
\title{make_segments_isdb}
\usage{
make_segments_isdb(
  isdb.df,
  do.qc = FALSE,
  return.choice = "lines",
  filename = NULL,
  createShp = TRUE,
  quietly = FALSE
)
}
\arguments{
\item{isdb.df}{This is the dataframe you want to plot.  isdb dataframes are
characterized by the existence of the following fields:
\itemize{
 \item \code{FISHSET_ID}
 \item \code{LAT1} and \code{LONG1}
 \item \code{LAT2} and \code{LONG2}
 \item \code{LAT3} and \code{LONG3}
 \item \code{LAT4} and \code{LONG4}
}}

\item{do.qc}{default is \code{FALSE}}

\item{return.choice}{This determines what the function returns.  The default
value of \code{"lines"} returns a SpatialLinesDataFrame, where the dataframe 
contains the generated QC information.  Any other value for 
\code{return.choice} will instead return of data frame of all of the 
FISHSET_IDs, the QC determination of each line, the length of each line (in 
km), and the number of valid vertices. 
The QC data checks the positions of each FISHSET for NAs, Zeroes, or 
otherwise data.  Additionally, if the line has a length of zero, that is 
noted as well.}

\item{filename}{default is \code{NULL}.  If you are outputting shapefiles, 
you can specify a name for them here.  They will also get a timestamp.}

\item{createShp}{default is \code{TRUE}. This determines whether or not 
shapefiles will be generated in your working directory.}

\item{quietly}{default is \code{FALSE}.  If TRUE, no output messages will be shown.}
}
\description{
This function takes a df of the isdb data (including fields
\code{FISHSET_ID} and LAT1:LAT4 and LONG1:LONG4) and makes a 
spatialLinesDataFrame which can be plotted.If selected, it can also QC them.
Each line will have the following columns in the resulting data frame:
\itemize{
 \item \code{FISHSET_ID} This uniquely identifies a set
 \item \code{QCPOS} This field identifies potential issues with the line - 
 including NA positions, positions that are "0", incorrect hemisphere, or 
 impossible coordinates
 \item \code{QCTIME} This field identifies potential issues with the timing 
 of the vertices making up the line.  For example,  P2 should occur after 
 P1, and 2 vertices cannot occur simultaneously.  Sets less than 5 min or 
 greater than 24 hr are also noted
 \item \code{LEN_KM} This field shows the calculated distance of the 
 resultant line in kms
 \item \code{N_VALID_VERT} This field shows how many vertices appear 
 correct after NAs, 0s and othe problematic values have been dropped
}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
