% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_fields.R
\name{rename_fields}
\alias{rename_fields}
\title{rename_fields}
\usage{
rename_fields(df = NULL, oldFields = NULL, newFields = NULL)
}
\arguments{
\item{df}{default is \code{NULL}. This is the dataframe you want to modify 
column names in.}

\item{oldFields}{default is \code{NULL}. This is a vector of fieldnames you 
want to rename.}

\item{newFields}{default is \code{NULL}. This is a vector of the new fieldnames 
you want to use.}
}
\value{
returns a data.frame with the new names.
}
\description{
This function facilitates the renaming of fields in a dataframe.
}
\note{
The number of items in oldFields and newFields must be identical, and 
the order matters in that newFields[3] will replace oldFields[3].
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
