% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{save_encrypted}
\alias{save_encrypted}
\title{.save_encrypted}
\usage{
save_encrypted(object = NULL, list = NULL, file, compress = TRUE)
}
\arguments{
\item{object}{R object to save}

\item{list}{Character vector of object names to save (NULL if using 'object')}

\item{file}{Path to save to}

\item{compress}{logical or character string specifying the compression method.}
}
\description{
saves data files such that they can't be easily loaded outside of the package
}
