% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{save_encrypted}
\alias{save_encrypted}
\title{save_encrypted}
\usage{
save_encrypted(
  object = NULL,
  list = NULL,
  file,
  envir = parent.frame(),
  encrypt = TRUE
)
}
\arguments{
\item{object}{R object to save}

\item{list}{Character vector of object names to save (NULL if using 'object')}

\item{file}{Path to save to}

\item{envir}{Environment to look in for objects (when using list)}

\item{encrypt}{Logical. If TRUE (default), performs encryption. If FALSE, saves normally via base::save.}
}
\description{
Saves data files encrypted or unencrypted based on a toggle such
that they can't be easily loaded by other users.
}
\seealso{
Other file_management: 
\code{\link{data.dir_valet}()},
\code{\link{load_encrypted}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{file_management}
