% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_in.r
\name{SQL_in}
\alias{SQL_in}
\title{SQL_in}
\usage{
SQL_in(x)
}
\arguments{
\item{x}{The default value is \code{NULL}.}
}
\description{
Converts a vector of characters into something that an "IN" statement can use
}
\seealso{
Other util: \code{\link{na_zero}}, \code{\link{st_err}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}l
}
