% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_err.r
\name{st_err}
\alias{st_err}
\title{st_err}
\usage{
st_err(x = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{The default value is \code{NULL}.}

\item{na.rm}{The default value is \code{FALSE}.}
}
\description{
Calculates the standard error
}
\seealso{
Other util: 
\code{\link{Mode}()},
\code{\link{SQL_in}()},
\code{\link{big_in}()},
\code{\link{combine_lists}()},
\code{\link{updateCheck}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{util}
