% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_err.r
\name{st_err}
\alias{st_err}
\title{st_err}
\usage{
st_err(x = NULL)
}
\arguments{
\item{x}{The default value is \code{NULL}.}
}
\description{
Calculates the standard error
}
\seealso{
Other util: 
\code{\link{SQL_in}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{util}
