% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCheck.R
\name{updateCheck}
\alias{updateCheck}
\title{updateCheck}
\usage{
updateCheck(gitPkg = NULL)
}
\arguments{
\item{gitPkg}{default is \code{NULL}. This is the URL to the DESCRIPTION file
on github.}
}
\description{
This function compares the package with the available version on 
github, and prompts the user to update.
}
\seealso{
Other general_use: \code{\link{df_qc_spatial}},
  \code{\link{df_to_sp}}, \code{\link{drop_NA_cols}},
  \code{\link{prepare_shape_fields}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
