% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VMS_clean_recs.R
\name{VMS_clean_recs}
\alias{VMS_clean_recs}
\title{VMS_clean_recs}
\usage{
VMS_clean_recs(df = NULL, lat.field = "LATITUDE",
  lon.field = "LONGITUDE", objField = "VR_NUMBER",
  timeField = "POSITION_UTC_DATE")
}
\arguments{
\item{df}{default is \code{NULL}.  This is the dataframe to be processed.  It 
should have coordinates in decimal degrees and they should be in fields 
called "LATITUDE" and "LONGITUDE".  It also needs a field with the time 
associated with each position, i.e. the  \code{timeField}.}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values 
(in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude 
values (in decimal degrees)}

\item{objField}{default is \code{VR_NUMBER}. This is a field identifying which 
points are associated withe eachother.}

\item{timeField}{default is \code{POSITION_UTC_DATE}. This is the field which 
will be used to calculate the time differences between records.}
}
\description{
This function takes raw VMS data (i.e. a dataframe having 
sequential coordinates and times), and removes duplicate positions.
}
\note{
Your resultant df will likely have less records than your source df.
During the cleaning process, values for Latitude and Longitude will be 
rounded to 4 decimals (~11m resolution), and values in your timeFields will be rounded 
to the nearest 5 minutes.  If a field called "UPDATE_DATE" exists, it will be 
used to grab only the most recent record in cases where multiple records have
the same values for both objField and timeField (e.g. vessel, and time of position)
}
\seealso{
Other vms: \code{\link{VMS_get_recs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{vms}
