% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VMS_from_MARFIS.R
\name{VMS_from_MARFIS}
\alias{VMS_from_MARFIS}
\title{VMS_from_MARFIS}
\usage{
VMS_from_MARFIS(
  df = NULL,
  cxn = NULL,
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  usepkg = "rodbc",
  data.dir = NULL,
  VR_field = "VR_NUMBER",
  LIC_field = "LICENCE_ID",
  GC_field = "GEAR_CODE",
  LANDED_field = "LANDED_DATE",
  look_ahead_days = 7,
  lat.field = NULL,
  lon.field = NULL,
  make_segments = TRUE,
  make_segments_spatial = FALSE
)
}
\arguments{
\item{df}{a dataframe to be analyzed.}

\item{cxn}{A valid Oracle connection object. This parameter allows you to 
pass an existing connection, reducing the need to establish a new connection 
within the function. If provided, it takes precedence over the connection-
related parameters.}

\item{fn.oracle.username}{Default is \code{'_none_'}. This is your username 
for accessing Oracle objects. If you have a value for \code{oracle.username} 
stored in your environment (e.g., from an rprofile file), this can be left 
out and that value will be used. If a value for this is provided, it will 
take priority over your existing value. Deprecated; use \code{cxn} instead.}

\item{fn.oracle.password}{Default is \code{'_none_'}. This is your password 
for accessing Oracle objects. If you have a value for \code{oracle.password} 
stored in your environment (e.g., from an rprofile file), this can be left 
out and that value will be used. If a value for this is provided, it will 
take priority over your existing value. Deprecated; use \code{cxn} instead.}

\item{fn.oracle.dsn}{Default is \code{'_none_'}. This is your DSN/ODBC 
identifier for accessing Oracle objects. If you have a value 
for \code{oracle.dsn} stored in your environment (e.g., from an rprofile 
file), this can be left out and that value will be used. If a value for this 
is provided, it will take priority over your existing value. Deprecated; use 
\code{cxn} instead.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the 
connection to Oracle should use \code{'rodbc'} or \code{'roracle'} to 
connect.  rodbc is slightly easier to setup, but roracle will extract data ~ 
5x faster. Deprecated; use \code{cxn} instead.}

\item{data.dir}{the default is \code{NULL}. If you are hoping to load existing data,
this folder should contain a data folder containing your rdata files. For this function, only
MARFLEETS_LIC will be used, and only to add gear and licenced species onto each record.}

\item{VR_field}{the default is \code{"VR_NUMBER"}}

\item{LIC_field}{the default is \code{"LICENCE_ID"}}

\item{GC_field}{the default is \code{"GEAR_CODE"}}

\item{LANDED_field}{the default is \code{"LANDED_DATE"}}

\item{look_ahead_days}{the default is \code{7}}

\item{lat.field}{the default is \code{NULL}. This is the name of the field holding latitude values 
(in decimal degrees)}

\item{lon.field}{the default is \code{NULL}. This is the name of the field holding longitude 
values (in decimal degrees)}

\item{make_segments}{the default is \code{TRUE}. This indicates whether or not the results should
include an sp object of all of the VMS data as lines (in addition to a data frame)}

\item{make_segments_spatial}{the default is \code{FALSE}. This indicates whether or not a shapefile
should be created of all of the VMS data as lines (in addition to a data frame). 
\code{make_segments} must be TRUE for shapefiles to be generated.}
}
\value{
a list containing a data.frame called "marf_VMS" of the joined marfis/VMS data, and, 
if \code{make_segments} is TRUE, an sp object called "marf_VMS_segments". Additionally, if 
make_segments and make_segments_shp are both TRUE, a shapefile will be created in the working 
directory.
}
\description{
This function takes a dataframe (with coordinate fields in decimal
degrees), and extracts asssociated VMS data
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
