% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VMS_from_MARFIS.R
\name{VMS_from_MARFIS}
\alias{VMS_from_MARFIS}
\title{VMS_from_MARFIS}
\usage{
VMS_from_MARFIS(
  df = NULL,
  cxn = NULL,
  data.dir = NULL,
  VR_field = "VR_NUMBER",
  LIC_field = "LICENCE_ID",
  GC_field = "GEAR_CODE",
  LANDED_field = "LANDED_DATE",
  look_ahead_days = 7,
  lat.field = NULL,
  lon.field = NULL,
  make_segments = TRUE,
  make_segments_spatial = FALSE,
  ...
)
}
\arguments{
\item{df}{a dataframe to be analyzed.}

\item{cxn}{A valid Oracle connection object. This parameter allows you to 
pass an existing connection, reducing the need to establish a new connection 
within the function.}

\item{data.dir}{the default is \code{NULL}. If you are hoping to load existing data,
this folder should contain a data folder containing your rdata files. For this function, only
MARFLEETS_LIC will be used, and only to add gear and licenced species onto each record.}

\item{VR_field}{the default is \code{"VR_NUMBER"}}

\item{LIC_field}{the default is \code{"LICENCE_ID"}}

\item{GC_field}{the default is \code{"GEAR_CODE"}}

\item{LANDED_field}{the default is \code{"LANDED_DATE"}}

\item{look_ahead_days}{the default is \code{7}}

\item{lat.field}{the default is \code{NULL}. This is the name of the field holding latitude values 
(in decimal degrees)}

\item{lon.field}{the default is \code{NULL}. This is the name of the field holding longitude 
values (in decimal degrees)}

\item{make_segments}{the default is \code{TRUE}. This indicates whether or not the results should
include an sp object of all of the VMS data as lines (in addition to a data frame)}

\item{make_segments_spatial}{the default is \code{FALSE}. This indicates whether or not a shapefile
should be created of all of the VMS data as lines (in addition to a data frame). 
\code{make_segments} must be TRUE for shapefiles to be generated.}

\item{...}{other arguments passed to methods - perhaps \code{extract_user} and \code{extract_computer}}
}
\value{
a list containing a data.frame called "marf_VMS" of the joined marfis/VMS data, and, 
if \code{make_segments} is TRUE, an sp object called "marf_VMS_segments". Additionally, if 
make_segments and make_segments_shp are both TRUE, a shapefile will be created in the working 
directory.
}
\description{
This function takes a dataframe (with coordinate fields in decimal
degrees), and extracts asssociated VMS data
}
\seealso{
Other vms: 
\code{\link{VMS_clean_recs}()},
\code{\link{VMS_get_recs}()},
\code{\link{subset_by_time}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{vms}
