% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VMS_get_recs.R
\name{VMS_get_recs}
\alias{VMS_get_recs}
\title{VMS_get_recs}
\usage{
VMS_get_recs(
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  usepkg = "rodbc",
  dateStart = NULL,
  dateEnd = NULL,
  vrnList = NULL,
  hrBuffer = 4,
  shp = NULL,
  shp.field = NULL,
  simpleQC = TRUE,
  rowNum = 50000,
  quietly = F
)
}
\arguments{
\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username} 
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}  
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for 
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file), 
this can be left and that value will be used.  If a value for this is 
provided, it will take priority over your existing value.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the 
connection to Oracle should use \code{'rodbc'} or \code{'roracle'} to 
connect.  rodbc is slightly easier to setup, but roracle will extract data 
~ 5x faster.}

\item{dateStart}{default is \code{NULL}. This is the start date ('YYYY-MM-DD') 
for your VMS data search}

\item{dateEnd}{default is \code{NULL}. This is the end date ('YYYY-MM-DD') 
for your VMS data search. If this is left NULL, one year of data will be
returned}

\item{vrnList}{default is \code{NULL}.  A vector or VRNs can be added so 
that the only positions returned match particular vessel(s).}

\item{hrBuffer}{default is \code{4}.  This is the number of hours worth of 
VMS data you would like to pad your area search by.  For example, if a vessel
has a single VMS position in the results, padding it will add additional 
points before and after the intrusion which can illustrate what it was doing 
at the time.}

\item{shp}{default is \code{NULL}.  This is the shapefile that has 
polygons that should be checked for existence of VMS data.}

\item{shp.field}{default is \code{NULL}.  This is the field in the shapefile
that should attached to the reurned data.  This will typically be a field 
that identifies the area(s).}

\item{simpleQC}{default is \code{TRUE}. If True, this drops records where the 
latitude is 0, 90 or -90 or the longitude is 0, 180 or -180.}

\item{rowNum}{default is \code{50000}.  This is the maximimum number of VMS 
records that can be extracted. It is in place to prevent crashing your 
application.}

\item{quietly}{default is \code{FALSE}.  This indicates whether or not
information about the process should be shown.}
}
\value{
a DataFrame with the column \code{agg.poly.field} added (if value for 
\code{shp} is supplied)
}
\description{
This function extracts VMS data for a given timespan and area.  
A time buffer can be added returns other points that are not within the area
of interest, but give context to the path.
}
\seealso{
Other vms: 
\code{\link{VMS_clean_recs}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{vms}
