% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VMS_get_recs.R
\name{VMS_get_recs}
\alias{VMS_get_recs}
\title{VMS_get_recs}
\usage{
VMS_get_recs(
  cxn = NULL,
  dateStart = NULL,
  dateEnd = NULL,
  vrnList = NULL,
  hrBuffer = 4,
  minLon = NULL,
  maxLon = NULL,
  minLat = NULL,
  maxLat = NULL,
  rowNum = 50000,
  quietly = F
)
}
\arguments{
\item{cxn}{A valid Oracle connection object. This parameter allows you to 
pass an existing connection, reducing the need to establish a new connection 
within the function. If provided, it takes precedence over the connection-
related parameters.}

\item{dateStart}{default is \code{NULL}. This is the start date ('YYYY-MM-DD') 
for your VMS data search}

\item{dateEnd}{default is \code{NULL}. This is the end date ('YYYY-MM-DD') 
for your VMS data search. If this is left NULL, one year of data will be
returned}

\item{vrnList}{default is \code{NULL}.  A vector or VRNs can be added so 
that the only positions returned match particular vessel(s).}

\item{hrBuffer}{default is \code{4}.  This is the number of hours worth of 
VMS data you would like to pad your area search by.  For example, if a vessel
has a single VMS position in the results, padding it will add additional 
points before and after the intrusion which can illustrate what it was doing 
at the time.}

\item{minLon}{default is \code{NULL}. This can be used to set a minimum Longitude}

\item{maxLon}{default is \code{NULL}. This can be used to set a maximum Longitude}

\item{minLat}{default is \code{NULL}. This can be used to set a minimum Latitude}

\item{maxLat}{default is \code{NULL}. This can be used to set a maximum Latitude}

\item{rowNum}{default is \code{50000}.  This is the maximum number of VMS 
records that can be extracted. It is in place to prevent crashing your 
application.}

\item{quietly}{default is \code{FALSE}.  This indicates whether or not
information about the process should be shown.}
}
\value{
a dataFrame
}
\description{
This function extracts VMS data for a given timespan and area.  
A time buffer can be added returns other points that are not within the area
of interest, but give context to the path.
}
\seealso{
Other vms: 
\code{\link{VMS_clean_recs}()},
\code{\link{VMS_from_MARFIS}()},
\code{\link{subset_by_time}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{vms}
