% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_now.R
\name{where_now}
\alias{where_now}
\title{where_now}
\usage{
where_now(inf = NULL, lvl = 1)
}
\arguments{
\item{inf}{the default is \code{NULL}.  WOrks with \code{as.character(sys.calls()[[sys.nframe() - 1]])}. 
This is a method for extracting the function information.}

\item{lvl}{the default is \code{1}. This controls how many tabs are used to 
display the results.  It can be useful when functions call subfunctions}
}
\value{
NULL - it just prints to the screen
}
\description{
This function indicates which function has been reached, and the 
parameters it is using
}
\seealso{
Other general_use: 
\code{\link{clip_by_poly_generic}()},
\code{\link{combine_lists}()},
\code{\link{df_qc_spatial}()},
\code{\link{df_to_shp}()},
\code{\link{df_to_sp}()},
\code{\link{drop_cols}()},
\code{\link{pbs2sp_poly}()},
\code{\link{prepare_shape_fields}()},
\code{\link{updateCheck}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
