% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_now.R
\name{where_now}
\alias{where_now}
\title{where_now}
\usage{
where_now(..., callstack = sys.calls())
}
\value{
nothing, it just writes messages to the screen
}
\description{
This function indicates which function has been reached, and the 
parameters it is using
}
\seealso{
Other debugging: 
\code{\link{changeDetector}()},
\code{\link{updateExpected}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{debugging}
