% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_now.R
\name{where_now}
\alias{where_now}
\title{where_now}
\usage{
where_now(returnTime = FALSE, callstack = sys.calls())
}
\arguments{
\item{returnTime}{default is \code{FALSE}.  If TRUE, this function will return a value for 
proc.time, facilitating timing of functions}

\item{callstack}{default is \code{sys.calls()}.  Not meant to be changed.}
}
\value{
nothing, it just writes messages to the screen
}
\description{
This function indicates which function has been reached.
}
\seealso{
Other debugging: 
\code{\link{changeDetector}()},
\code{\link{updateExpected}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{debugging}
